WE-3211: rename model
