+++ src/models/wasteTreatmentCapacity.model.js
@@ -1,172 +1,119 @@
 'use strict'
 
-const DataStore = require('../models/dataStore.model')
-const ApplicationLine = require('../persistence/entities/applicationLine.entity')
-const ApplicationAnswer = require('../persistence/entities/applicationAnswer.entity')
+const DataStore =
+  require('../models/dataStore.model')
+const ApplicationLine =
+  require('../persistence/entities/applicationLine.entity')
+const ApplicationAnswer =
+  require('../persistence/entities/applicationAnswer.entity')
 
-const TREATMENT_HAZARDOUS_ANSWER = 'treatment-hazardous'
-const TREATMENT_BIOLOGICAL_ANSWER = 'treatment-biological'
-const TREATMENT_CHEMICAL_ANSWER = 'treatment-chemical'
-const TREATMENT_INCINERATION_ANSWER = 'treatment-incineration'
-const TREATMENT_SLAGS_ANSWER = 'treatment-slags'
-const TREATMENT_METAL_ANSWER = 'treatment-metal'
-const TREATMENT_NONE_ANSWER = 'treatment-none'
+const treatmentAnswers = [
+  {
+    questionCode: 'treatment-hazardous',
+    questionText: 'Treatment of hazardous waste',
+    answerCode: 'no'
 
-const TREATMENT_HAZARDOUS_DAILY = 'treatment-hazardous-daily'
-const TREATMENT_BIOLOGICAL_DAILY = 'treatment-biological-daily'
-const TREATMENT_CHEMICAL_DAILY = 'treatment-chemical-daily'
-const TREATMENT_INCINERARION_DAILY = 'treatment-incineration-daily'
-const TREATMENT_SLAGS_DAILY = 'treatment-slags-daily'
-const TREATMENT_METAL_DAILY = 'treatment-metal-daily'
+  },
+  {
+    questionCode: 'treatment-biological',
+    questionText: 'Biological treatment (eg composting or anaerobic digestion)',
+    answerCode: 'no'
 
-const APPLICATION_ANSWERS = {
-  [TREATMENT_HAZARDOUS_ANSWER]: 'hazAnswer',
-  [TREATMENT_BIOLOGICAL_ANSWER]: 'bioAnswer',
-  [TREATMENT_CHEMICAL_ANSWER]: 'chemAnswer',
-  [TREATMENT_INCINERATION_ANSWER]: 'incinerationAnswer',
-  [TREATMENT_SLAGS_ANSWER]: 'slagAnswer',
-  [TREATMENT_METAL_ANSWER]: 'metalAnswer',
-  [TREATMENT_NONE_ANSWER]: 'noneAnswer',
-  [TREATMENT_HAZARDOUS_DAILY]: 'hazWeight',
-  [TREATMENT_BIOLOGICAL_DAILY]: 'bioWeight',
-  [TREATMENT_CHEMICAL_DAILY]: 'chemWeight',
-  [TREATMENT_INCINERARION_DAILY]: 'incinerationWeight',
-  [TREATMENT_SLAGS_DAILY]: 'slagWeight',
-  [TREATMENT_METAL_DAILY]: 'metalWeight'
-}
+  },
+  {
+    questionCode: 'treatment-chemical',
+    questionText: 'Physico-chemical treatment for disposal',
+    answerCode: 'no'
 
-async function listForWasteActivitiesMinusDiscountLines (context) {
+  },
+  {
+    questionCode: 'treatment-incineration',
+    questionText: 'Pre-treatment of waste for incineration or co-incineration',
+    answerCode: 'no'
+
+  },
+  {
+    questionCode: 'treatment-slags',
+    questionText: 'Treatment of slags and ashes',
+    answerCode: 'no'
+
+  },
+  {
+    questionCode: 'treatment-metal',
+    questionText: `Treatment in shredders of metal waste, including
+      waste electrical and electronic equipment and end-of-life
+      vehicles and their components`,
+    answerCode: 'no'
+
+  },
+  {
+    questionCode: 'treatment-none',
+    questionText: 'None of the above',
+    answerCode: 'no'
+  }
+]
+
+async function getRelevantApplicationLine (context, activityIndex) {
   // fetch all application lines, including discounts
-  const allWasteTreatmentCapacityApplicationLines = await ApplicationLine.listForWasteActivities(context)
+  const allWasteTreatmentCapacityApplicationLines =
+    await ApplicationLine.listForWasteActivities(context)
   // filter out the discounts, prevents them being included in task list
-  const wasteTreatmentCapacityApplicationLines = allWasteTreatmentCapacityApplicationLines.filter(({ value }) => value > -1)
-  return wasteTreatmentCapacityApplicationLines
+  const wasteTreatmentCapacityApplicationLines =
+    allWasteTreatmentCapacityApplicationLines.filter(({ value }) => value > -1)
+  return {
+    applicationLine: wasteTreatmentCapacityApplicationLines[activityIndex],
+    hasNext: wasteTreatmentCapacityApplicationLines
+      .slice(activityIndex + 1)
+      .length > 0
+  }
 }
 
-module.exports = class WasteTreatmentCapacities {
-  constructor ({
-    forActivityIndex = 0,
-    activityDisplayName = '',
-    hasNext = false,
-    hazAnswer,
-    bioAnswer,
-    chemAnswer,
-    incinerationAnswer,
-    slagAnswer,
-    metalAnswer,
-    noneAnswer,
-    hazWeight,
-    bioWeight,
-    chemWeight,
-    incinerationWeight,
-    slagWeight,
-    metalWeight
-  } = {}) {
-    Object.assign(this, {
-      forActivityIndex,
-      activityDisplayName,
-      hasNext,
-      hazAnswer,
-      bioAnswer,
-      chemAnswer,
-      incinerationAnswer,
-      slagAnswer,
-      metalAnswer,
-      noneAnswer,
-      hazWeight,
-      bioWeight,
-      chemWeight,
-      incinerationWeight,
-      slagWeight,
-      metalWeight
-    })
-  }
+async function saveAnswer (answer, context) {
+  const saver = new ApplicationAnswer(answer)
+  return saver.save(context)
+}
 
-  static async getForActivity (context, activityIndex) {
-    const wasteTreatmentCapacityApplicationLines = await listForWasteActivitiesMinusDiscountLines(context)
-    const wasteTreatmentCapacityApplicationLine = wasteTreatmentCapacityApplicationLines[activityIndex]
+module.exports = {
+  treatmentAnswers,
+  getForActivity: async function (context, activityIndex) {
+    const { applicationLine, hasNext } =
+      await getRelevantApplicationLine(context, activityIndex)
 
-    if (wasteTreatmentCapacityApplicationLine) {
-      const hasNext = Boolean(wasteTreatmentCapacityApplicationLines[activityIndex + 1])
-      const definedName = wasteTreatmentCapacityApplicationLine.lineName || ''
-      const activityDisplayName = `${definedName}`.trim()
-      const { data: { acceptsHazardousWaste } } = await DataStore.get(context)
-      const hasHazardousWaste = Boolean(acceptsHazardousWaste)
-
-      const wasteTreatmentCapacityAnswers = await ApplicationAnswer
-        .listForApplicationLine(
-          context,
-          wasteTreatmentCapacityApplicationLine.id,
-          Object.keys(APPLICATION_ANSWERS)
-        )
-      console.log(wasteTreatmentCapacityAnswers)
-      const data = wasteTreatmentCapacityAnswers.reduce((acc, { questionCode, answerText }) => {
-        const propertyName = APPLICATION_ANSWERS[questionCode]
-        acc[propertyName] = answerText
-        return acc
-      }, {})
-      const { nonHazardousThroughput, nonHazardousMaximum, hazardousThroughput, hazardousMaximum } = data
-
-      return new WasteTreatmentCapacities({ forActivityIndex: activityIndex, activityDisplayName, hasNext, hasHazardousWaste, nonHazardousThroughput, nonHazardousMaximum, hazardousThroughput, hazardousMaximum })
-    }
-  }
-
-  static async getAllForApplication (context) {
-    const wasteTreatmentCapacityApplicationLines = await ApplicationLine.listForWasteActivities(context)
-    if (!wasteTreatmentCapacityApplicationLines) {
-      return []
-    }
-
-    const { data: { acceptsHazardousWaste } } = await DataStore.get(context)
-    const hasHazardousWaste = Boolean(acceptsHazardousWaste)
-
-    const allWasteWeightAnswers = await ApplicationAnswer
-      .listByMultipleQuestionCodes(context, Object.keys(APPLICATION_ANSWERS))
-
-    return wasteTreatmentCapacityApplicationLines.map((wasteTreatmentCapacityApplicationLine, index, allLines) => {
-      const hasNext = Boolean(allLines[index + 1])
-      const definedName = wasteTreatmentCapacityApplicationLine.lineName || ''
-      const activityDisplayName = `${definedName}`.trim()
-
-      const wasteTreatmentCapacityAnswers = allWasteWeightAnswers.filter(({ applicationLineId }) => applicationLineId === wasteTreatmentCapacityApplicationLine.id)
-
-      const data = wasteTreatmentCapacityAnswers.reduce((acc, { questionCode, answerText }) => {
-        const propertyName = APPLICATION_ANSWERS[questionCode]
-        acc[propertyName] = answerText
-        return acc
-      }, {})
-      const { nonHazardousThroughput, nonHazardousMaximum, hazardousThroughput, hazardousMaximum } = data
-
-      return new WasteTreatmentCapacities({ forActivityIndex: index, activityDisplayName, hasNext, hasHazardousWaste, nonHazardousThroughput, nonHazardousMaximum, hazardousThroughput, hazardousMaximum })
-    })
-  }
-
-  async save (context) {
-    const wasteTreatmentCapacityApplicationLines = await listForWasteActivitiesMinusDiscountLines(context)
-    const wasteTreatmentCapacityApplicationLine = wasteTreatmentCapacityApplicationLines[this.forActivityIndex]
-    if (wasteTreatmentCapacityApplicationLine) {
-      const applicationLineId = wasteTreatmentCapacityApplicationLine.id
-      const applicationAnswerSaves = []
-      applicationAnswerSaves.push(new ApplicationAnswer({
-        applicationLineId,
-        questionCode: TREATMENT_HAZARDOUS_ANSWER,
-        answerText: this.nonHazardousThroughput
-      }).save(context))
-
-      await Promise.all(applicationAnswerSaves)
-    }
-  }
-
-  static async getAllWeightsHaveBeenEnteredForApplication (context) {
-    const wasteTreatmentCapacityApplicationLines = await listForWasteActivitiesMinusDiscountLines(context)
-    const requiredEntries = []
+    const activityDisplayName =
+      (applicationLine.lineName || '').trim()
 
     const wasteTreatmentCapacityAnswers = await ApplicationAnswer
-      .listByMultipleQuestionCodes(context, Object.keys(APPLICATION_ANSWERS))
+      .listForApplicationLine(
+        context,
+        applicationLine.id,
+        treatmentAnswers.map(answer => answer.questionCode)
+      )
+    return {
+      wasteTreatmentCapacityAnswers,
+      activityDisplayName,
+      hasNext
+    }
+  },
+  saveAnswers: async function (context, activityIndex, answers) {
+    const positiveAnswers = Object.keys(answers)
+    const { applicationLine, hasNext } =
+      await getRelevantApplicationLine(context, activityIndex)
 
-    return wasteTreatmentCapacityApplicationLines
-      .every(({ id }) => requiredEntries
-        .every((requiredEntry) => wasteTreatmentCapacityAnswers
-          .find(({ applicationLineId, questionCode, answerText }) => (applicationLineId === id) && (questionCode === requiredEntry) && answerText)))
+    const toSave = treatmentAnswers.map(ta => ({
+      applicationLineId: applicationLine.id,
+      questionCode: ta.questionCode,
+      answerCode: positiveAnswers.indexOf(ta.questionCode) >= 0 ? 'yes' : 'no'
+    }))
+
+    const savePromises = toSave.map(answer => saveAnswer(answer, context))
+    await Promise.all(savePromises)
+
+    return {
+      hasNext,
+      forActivityIndex: activityIndex
+    }
+  },
+  getAllWeightsHaveBeenEnteredForApplication: async function (context) {
+    return []
   }
 }
